feat: Implement BlackBox style injection for Android 16-18
Refactors injection logic to support Android 16-18 (API 36-38) using a BlackBox-style isolation approach. This includes updates to VirtualContainer and VirtualSpaceInjector to handle newer Android restrictions and sandboxing mechanisms.

The UniversalInjector is simplified to delegate directly to the new LibraryInjector, which centralizes the version-specific injection strategies.

The OffsetFinder utility is adjusted to correctly parse patterns for AOB scanning.

+++ loader/src/main/java/com/onecore/loader/OffsetFinder.java
@@ -72,7 +72,7 @@ public class OffsetFinder {
      * Scans memory for a pattern of bytes (AOB Scan).
      */
     public static long findPattern(int pid, long start, long end, String pattern) {
-        byte[] patternBytes = parsePattern(pattern);
+        int[] patternBytes = parsePattern(pattern);
         int patternLen = patternBytes.length;
         
         // Optimized scanning logic (Searching in 1MB chunks to avoid OOM)
